requirements.txt
streamlit
python-dotenv
requests
trafilatura
edge-tts

groq

elevenlabs
pydub